gym/requirements.txt
asgiref==3.5.0
Django==3.2.12
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
Pillow==9.1.0
psycopg2==2.9.3
PyJWT==1.7.1
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.1.1
